fix(lotto): 기존 로또 클래스 변경
Lotto 클래스를 PlayLotto 클래스로 변경
Lotto 클래스 재구성
