+++ src/test/java/lotto/LottoTest.java
@@ -5,9 +5,11 @@ import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.RepetitionInfo;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
 
@@ -15,27 +17,27 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 public class LottoTest {
 
-    private Lotto lotto;
+    private PlayLotto playLotto;
 
     @BeforeEach
     void setUp() {
-        lotto = new Lotto();
+        playLotto = new PlayLotto();
     }
 
     @ParameterizedTest
     @ValueSource(ints = {14000})
     void buyLotto(int money) {
-        assertThat(lotto.buyLotto(money)).isEqualTo(14);
+        assertThat(playLotto.buyLotto(money)).isEqualTo(14);
     }
 
     @Test
     void getRandomNumber(){
-        assertThat(lotto.getRandomNumber()).isBetween(1, 45);
+        assertThat(playLotto.getRandomNumber()).isBetween(1, 45);
     }
 
     @RepeatedTest(value = 6)
     void createLotto(RepetitionInfo repetitionInfo){
-        assertThat(lotto.createLotto().get(repetitionInfo.getCurrentRepetition() - 1)).isBetween(1, 45);
+        assertThat(playLotto.createLotto().get(repetitionInfo.getCurrentRepetition() - 1)).isBetween(1, 45);
     }
 
 }
